requirements.txt
mininet==2.3.0.dev6
PyYAML==6.0.1
setuptools==69.2.0
